fix: directory fix
