+++ conf/requirements.txt
@@ -2,3 +2,35 @@ tensorflow==2.2.0
 numpy==1.17.4
 torch==1.7.1
 PyYAML==5.4.1
+torch==1.7.1
+tensorflow==2.2.0
+numpy==1.17.4
+PyYAML==5.4.1
+numpy==1.17.4
+torch==1.7.1
+tensorflow==2.2.0
+PyYAML==5.4.1
+numpy==1.17.4
+tensorflow==2.2.0
+torch==1.7.1
+PyYAML==5.4.1
+numpy==1.17.4
+tensorflow==2.2.0
+torch==1.7.1
+PyYAML==5.4.1
+torch==1.7.1
+tensorflow==2.2.0
+numpy==1.17.4
+PyYAML==5.4.1
+numpy==1.17.4
+tensorflow==2.2.0
+torch==1.7.1
+PyYAML==5.4.1
+numpy==1.17.4
+torch==1.7.1
+tensorflow==2.2.0
+PyYAML==5.4.1
+numpy==1.17.4
+torch==1.7.1
+tensorflow==2.2.0
+PyYAML==5.4.1
