+++ src/test.py
@@ -14,7 +14,7 @@ class Test(unittest.TestCase):
 
     def test_case1(self):
         if self.test_configs["input_type"] == "text":
-            r = {"input": "안녕 세계"}
+            r = {"input": ["Hello worlds", "Hi worlds"]}
         else:
             dim = self.test_configs["input_dimension"]
             r = {"input": np.ndarray(dim).tolist()}
@@ -25,7 +25,7 @@ class Test(unittest.TestCase):
 
     def test_case2(self):
         if self.test_configs["input_type"] == "text":
-            r = {"input": "다시 안녕 세계"}
+            r = {"input": ["Hello again worlds", "Hi again worlds"]}
         else:
             dim = self.test_configs["input_dimension"]
             r = {"input": np.ndarray(dim).tolist()}
